Engie/requirements.txt
django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.10.0
